@@ -21,6 +21,6 @@ class NewRelicController < ApplicationController
 
   def ensure_new_reclic_api_key
     return if SamsonNewRelic::Api::KEY
-    head text: "NewReclic api key is not configured", status: :precondition_failed
+    render plain: "NewReclic api key is not configured", status: :precondition_failed
   end
 end
